Extract createUpdateEndowmentControllerMsg
