+++ src/pages/Admin/Charity/Permissions/useSubmit.ts
@@ -1,9 +1,5 @@
 import { useFormContext } from "react-hook-form";
 import { ProposalMeta } from "pages/Admin/types";
-import {
-  SettingsPermissions,
-  UpdateEndowmentControllerMsg,
-} from "types/contracts";
 import { useAdminResources } from "pages/Admin/Guard";
 import { useErrorContext } from "contexts/ErrorContext";
 import { useGetWallet } from "contexts/WalletContext";
@@ -12,7 +8,8 @@ import SettingsController from "contracts/SettingsController";
 import useCosmosTxSender from "hooks/useCosmosTxSender";
 import { isEmpty } from "helpers";
 import { getPayloadDiff } from "helpers/admin";
-import { FormField, FormValues } from "./schema";
+import createUpdateEndowmentControllerMsg from "./createUpdateEndowmentControllerMsg";
+import { FormValues } from "./schema";
 
 export default function useSubmit() {
   const { id, cw3, propMeta, endow_type } = useAdminResources<"charity">();
@@ -62,52 +59,4 @@ export default function useSubmit() {
     reset,
     submit: handleSubmit(onSubmit),
   };
-}
-
-function createUpdateEndowmentControllerMsg(
-  endowId: number,
-  formValues: Partial<Omit<FormValues, "initialValues">>
-): UpdateEndowmentControllerMsg {
-  const accountFees = createField(formValues.accountFees);
-  const beneficiaries_allowlist = createField(formValues.accountFees);
-  const donationSplitParams = createField(formValues.donationSplitParams);
-  const profile = createField(formValues.profile);
-  const contributors_allowlist = createField(formValues.contributors_allowlist);
-
-  const updateMsg: UpdateEndowmentControllerMsg = {
-    id: endowId,
-    aum_fee: accountFees,
-    beneficiaries_allowlist,
-    categories: profile,
-    contributors_allowlist,
-    deposit_fee: accountFees,
-    earnings_fee: accountFees,
-    ignore_user_splits: donationSplitParams,
-    image: profile,
-    kyc_donors_only: profile,
-    logo: profile,
-    maturity_allowlist: beneficiaries_allowlist,
-    name: profile,
-    split_to_liquid: donationSplitParams,
-    withdraw_fee: accountFees,
-  };
-  return updateMsg;
-}
-
-function createField(formField?: FormField): SettingsPermissions | undefined {
-  if (!formField) {
-    return undefined;
-  }
-
-  const result: SettingsPermissions = {
-    gov_controlled: formField.gov_controlled,
-    modifiable: formField.modifiable,
-    owner_controlled: formField.owner_controlled,
-  };
-
-  if (formField.delegate) {
-    result.delegate = { address: formField.delegate_address };
-  }
-
-  return result;
 }
